Major - Questionnaire editor position changes and callback fixes
