+++ src/components/Footer/index.tsx
@@ -1,7 +1,7 @@
 import { DefaultFooter } from '@ant-design/pro-layout';
 
 const Footer: React.FC = () => {
-  const defaultMessage = 'iSee: Building the AI you trust - v1.013 Feb 08';
+  const defaultMessage = 'iSee: Building the AI you trust - v1.014 Feb 22';
   const currentYear = new Date().getFullYear();
   return (
     <DefaultFooter
