+++ src/components/iSee/question/toolkit/QuestionnaireEditor.tsx
@@ -1,7 +1,7 @@
-import { PlusCircleOutlined, HolderOutlined, ImportOutlined, } from '@ant-design/icons';
+import { PlusCircleOutlined, HolderOutlined, ImportOutlined, DownloadOutlined, PlusOutlined, DashboardOutlined, FieldStringOutlined, CheckCircleOutlined, CheckSquareOutlined, FieldNumberOutlined, } from '@ant-design/icons';
 import Draggable from 'react-draggable';
 import QuestionForm from '@/components/iSee/question/toolkit/QuestionForm';
-import { Button, Modal, notification } from 'antd';
+import { Button, Checkbox, Col, Form, message, Modal, notification, Row, Select, Space, Tooltip } from 'antd';
 import React, { useCallback, useState, useEffect } from 'react';
 
 import TextArea from 'antd/lib/input/TextArea';
@@ -13,6 +13,7 @@ import { v4 as uuidv4 } from 'uuid';
 import type { Question, Response } from '@/models/questionnaire';
 
 import useDrag from '@/components/iSee/question/toolkit/useDrag';
+import { api_get_all } from '@/services/isee/questionnaires';
 
 const QuestionnaireEditor: React.FC<{
   defaultQuestions?: Question[];
@@ -30,10 +31,29 @@ const QuestionnaireEditor: React.FC<{
     '1.5rem',
   );
 
+  const [modalQuestionnaires, setModalQuestionnaires] = useState({ visibility: false });
+  const [importQuestionnaire, setImportQuestionnaire] = useState(undefined);
+  const [importedQuestionnaires, setImportQuestionnaires] = useState<Questionnaire[]>([]);
+
   // useEffect(() => {
   //   console.log("RUNNNING HERE ", defaultQuestions)
   //   setQuestions(defaultQuestions || []);
   // }, [defaultQuestions, setQuestions]);
+
+  useEffect(() => {
+    (async () => {
+      const data = await api_get_all();
+      setImportQuestionnaires(data);
+    })();
+  }, []);
+
+  const questionTypeToIcon = {
+    Likert: <DashboardOutlined />,
+    'Free-Text': <FieldStringOutlined />,
+    Radio: <CheckCircleOutlined />,
+    Checkbox: <CheckSquareOutlined />,
+    Number: <FieldNumberOutlined />,
+  };
 
   const removeAttr = (list: Question[], attr: string) => {
     return list.map((elmnt: Question) => {
@@ -109,8 +129,27 @@ const QuestionnaireEditor: React.FC<{
     ]);
   };
 
+  const onFinishImportQuestionnaire = (values: any) => {
+    const questionList: Question[] =
+      // collect questions based on the id from pre-existing questionnaires
+      importedQuestionnaires
+        .filter((q) => q._id == values.questionnaire)[0]
+        ?.questions?.filter((q) => values.questions.includes(q.id))
+        ?.map((q) => ({
+          ...q,
+          id: 'q-' + uuidv4(),
+          dimension: importedQuestionnaires.filter((qs) => qs._id == values.questionnaire)[0].dimension,
+        })) || [];
+
+    handleCloseModalQuestionnaires();
+    setQuestions([...questions, ...questionList]);
+    message.success({ content: 'Succesfully Imported Questions', duration: 2 });
+  };
+
+
   const handleQuestionChange = useCallback(
     (newQuestion: Question) => {
+      console.log("useCallback INNER")
       setQuestions((old: Question[]) => {
         const index = old.findIndex((q: Question) => q.id === newQuestion.id);
         if (index === -1) return old;
@@ -140,8 +179,16 @@ const QuestionnaireEditor: React.FC<{
     setModal({ visibility: true });
   };
 
+  const handleImportQuestionaireButton = () => {
+    setModalQuestionnaires({ visibility: true });
+  };
+
   const handleCloseModal = () => {
     setModal({ ...modal, visibility: false });
+  };
+
+  const handleCloseModalQuestionnaires = () => {
+    setModalQuestionnaires({ ...modalQuestionnaires, visibility: false });
   };
 
   return (
@@ -182,7 +229,7 @@ const QuestionnaireEditor: React.FC<{
           onClick={handleAddbutton}
           className="add-question-button"
         >
-          Add a question
+          Create New Question
         </Button>
       )}
       {!noImport && questions.length === 0 && (
@@ -195,6 +242,28 @@ const QuestionnaireEditor: React.FC<{
           Import a questionnaire
         </Button>
       )}
+
+      <Row style={{ marginBottom: 20 }}>
+        <Col span={8} offset={9}>
+          <Space >
+            <Button
+              icon={<DownloadOutlined />}
+              // type="primary"
+              onClick={handleImportQuestionaireButton}
+              size='large'>
+              Import Questionnaire
+            </Button>
+            <Button
+              size='large'
+              onClick={handleAddbutton}
+              icon={<PlusOutlined />}
+            >
+              Create New
+            </Button>
+          </Space>
+        </Col>
+      </Row>
+
       <Modal
         title={'Import'}
         visible={modal.visibility}
@@ -217,6 +286,85 @@ const QuestionnaireEditor: React.FC<{
           </code>
         </pre>
       </Modal>
+
+      {/* Load Question Popup  */}
+      <Modal
+        title="Load Questionnaire"
+        visible={modalQuestionnaires.visibility}
+        onCancel={handleCloseModalQuestionnaires}
+        destroyOnClose={true}
+        footer={[
+          <Button key="back"
+            onClick={handleCloseModalQuestionnaires}
+          >
+            Cancel
+          </Button>,
+          <Button form="importQuestionForm" key="submitQ" htmlType="submit" type="primary">
+            Load Questions
+          </Button>,
+        ]}
+      >
+        <Form
+          id="importQuestionForm"
+          name="importQuestionForm"
+          layout="vertical"
+          preserve={false}
+          labelCol={{ span: 0 }}
+          onFinish={onFinishImportQuestionnaire}
+          autoComplete="off"
+          onValuesChange={(changedValues) =>
+            changedValues?.questionnaire
+              ? setImportQuestionnaire(changedValues.questionnaire)
+              : null
+          }
+        >
+          <Form.Item
+            shouldUpdate
+            label="Select a Questionnaire"
+            name="questionnaire"
+            tooltip="This is a required field"
+            initialValue={importQuestionnaire}
+            rules={[{ required: true, message: 'Input is required!' }]}
+          >
+            <Select defaultValue={importQuestionnaire}>
+              {importedQuestionnaires.map((importedQuestionnaire) => (
+                <Select.Option key={importedQuestionnaire._id} value={importedQuestionnaire._id}>
+                  {importedQuestionnaire.name}
+                </Select.Option>
+              ))}
+            </Select>
+          </Form.Item>
+          {importQuestionnaire && (
+            <Form.Item
+              className="question-select-checkbox"
+              label="Select Questions"
+              name="questions"
+              tooltip="This is a required field"
+              rules={[{ required: true, message: 'Input is required!' }]}
+            >
+              <Checkbox.Group
+                className="checkbox-group"
+                options={
+                  importedQuestionnaires
+                    .filter((q) => q._id == importQuestionnaire)[0]
+                    ?.questions?.map((q) => ({
+                      label: (
+                        <>
+                          <Tooltip title={`This is a ${q.responseType} question`}>
+                            {questionTypeToIcon[q.responseType ?? '']}
+                          </Tooltip>
+                          {q.content ?? ''}
+                        </>
+                      ),
+                      value: q.id ?? '',
+                    })) || []
+                }
+              />
+            </Form.Item>
+          )}
+        </Form>
+      </Modal>
+
     </>
   );
 };
